+++ RELEASES.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## 0.2.2
+
+### Fixes
+
+- Border or padding on the horizontal axis could, in some cases, increase the height of nodes.
+- Vertical `border` and `padding` percentage values incorrectly multiplied by `available_space.width` instead of `height`.
+
 ## 0.2.1
 
 ### Fixes